Fix declaration of card size
The `card` string is used in a `getcwd(card, SZ_PATHNAME)` call. The
man page or getcwd says however:

> char *getcwd(char *buf, size_t size);
> [...]
> The size argument is the size in bytes of the character array pointed
> to by the buf argument.

`SZ_PATHNAME` is set to 511; however `card` had only 256 chars. With
hardening enabled, this lead to a segmentation fault.

This patch just increases the size of `card` to the necessary value.