Remove (most of) cfitsio
Cfitsio is used only for the fpack/funpack executables. However, the
main IRAF package (>2.16.1) already provides the cfitsio
library (bin$libcfitsio.a), so that we don't need to recompile it
ourself.

We only provide the required additional sources for fpack and funpack,
together with their license and a simple mkpkg script. The sources
come from cfitsio version 3.450.